video-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
ffmpeg-python==0.2.0
aiofiles==23.2.1
python-multipart==0.0.6
pika==1.3.2
httpx==0.25.1
pydantic-settings==2.1.0
redis==5.0.1
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==5.11.0
celery-prometheus-exporter==1.7.0
